requirements.txt
# All packages except for PyTorch
stable-baselines3
torchattacks
gymnasium
pyyaml
timm
datasets
transformers